requirements.txt
Django==1.7
Pillow==2.6.1
argparse==1.2.1
wsgiref==0.1.2
